Cleaned up IMC code
